pyproject.toml
[project]
name = "all-datasets"
version = "0.1.0"
description = "Various datasets of all kinds"
readme = ".github/README.md"
requires-python = ">=3.12"

dependencies = [
    "beautifulsoup4>=4.12.3",
    "huggingface-hub>=0.26.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ruff>=0.7.1",
]
